Make craftedWildCardFilter use CSV and add popup (#570)
* Make craftedWildCardFilter use CSV and add popup

- Modified filterPatternName so it can accept comma separated values.
This will allow the user to enter specific names such as "THE CALL" or
multiple values such as "THE CALL,NO HESITATION". Note that the input
is always lowercased for comparisons.
- make the user interface include a popup that allows default values for
the craftedWildCardFilter. Also include a button to clear the field.
Make the text field a bit wider to more easily enter multiple values.
- populate craftedWildCardFilter popup menu with "current seasonal"
and "ghost" options.
- update package.json version number

* update version

---------

Co-authored-by: John Hunkins <30057902+jchunkins@users.noreply.github.com>